changed saw base frequency to float and moved init of vocoder to synthserv

+++ headers/utils.h
@@ -4,4 +4,4 @@
 float freqtoang (float freq, int fs);
 
 // saw data to period
-float getPeriod(int fs, int key, int f0, int k0, int numkeys);
+float getPeriod(int fs, int key, float f0, int k0, int numkeys);
